+++ client/src/components/Shows/ShowsPage.jsx
@@ -1,4 +1,4 @@
-import '../App.css'
+import '../../App.css'
 import Button from '@mui/material/Button';
 import Card from '@mui/material/Card';
 import CardActions from '@mui/material/CardActions';
